backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
Pillow>=10.0.0
browser-use>=0.1.40 